Carry loading config with catalog.
