+++ src/lsdb/__init__.py
@@ -3,7 +3,7 @@ from hats.pixel_math import HealpixPixel
 from ._version import __version__
 from .catalog import Catalog, MarginCatalog
 from .core.crossmatch.crossmatch import crossmatch
-from .core.search import BoxSearch, ConeSearch, PixelSearch, PolygonSearch
+from .core.search.region_search import BoxSearch, ConeSearch, PixelSearch, PolygonSearch
 from .loaders.dataframe.from_dataframe import from_dataframe
 from .loaders.hats.read_hats import open_catalog, read_hats
-from .nested.datasets import generate_catalog
+from .nested.datasets import generate_catalog, generate_data
