--- src/lsdb/core/search/__init__.py
@@ -1,7 +1,0 @@
-from .box_search import BoxSearch
-from .cone_search import ConeSearch
-from .index_search import IndexSearch
-from .moc_search import MOCSearch
-from .order_search import OrderSearch
-from .pixel_search import PixelSearch
-from .polygon_search import PolygonSearch
