--- src/lsdb/loaders/__init__.py
@@ -1,2 +1,0 @@
-from .dataframe import from_dataframe
-from .hats import open_catalog, read_hats
