--- src/lsdb/loaders/dataframe/__init__.py
@@ -1,1 +1,0 @@
-from .from_dataframe import from_dataframe
