--- src/lsdb/loaders/hats/__init__.py
@@ -1,1 +1,0 @@
-from .read_hats import open_catalog, read_hats
