+++ src/lsdb/nested/__init__.py
@@ -1,5 +1,4 @@
 from . import accessor, backends  # noqa
 from .core import NestedFrame  # noqa
-from .datasets import generate_catalog, generate_data  # noqa
 from .io import read_parquet  # noqa
 from .utils import count_nested  # noqa
