+++ tests/lsdb/catalog/test_index_search.py
@@ -7,7 +7,7 @@ from hats.catalog.dataset.collection_properties import CollectionProperties
 
 import lsdb
 import lsdb.nested as nd
-from lsdb.core.search import IndexSearch
+from lsdb.core.search.index_search import IndexSearch
 
 
 def test_id_search_with_single_field(small_sky_order1_source_collection_catalog, helpers):
