+++ tests/lsdb/catalog/test_order_search.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 
 import lsdb.nested as nd
-from lsdb.core.search import OrderSearch
+from lsdb.core.search.region_search import OrderSearch
 
 
 def test_order_search_filters_correct_pixels(small_sky_source_catalog, helpers):
