+++ tests/lsdb/catalog/test_pixel_search.py
@@ -3,7 +3,7 @@ import pytest
 from hats.pixel_math import HealpixPixel
 
 import lsdb.nested as nd
-from lsdb.core.search.pixel_search import PixelSearch
+from lsdb.core.search.region_search import PixelSearch
 
 
 def test_pixel_search(small_sky_catalog, small_sky_order1_catalog):
